pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=0.3.23",
    "faiss-cpu>=1.13.0",
    "ipykernel>=7.1.0",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.7",
    "langchain-groq>=1.0.1",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "pymupdf>=1.26.6",
    "pypdf>=6.3.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "typesense>=1.3.0",
]
